[SDK] First version of the SDK
On this commit we have the basic bluetooth connectivity to interact with
the Clip.

Connectivity events, calibration commands, data capturing and example of
FlickGym wave to represent the data.

There are three main classes:

FlicktekSettings storing all the settings and letting Application or
wear to share data.

FlicktekManager controlling the state of the clip

FlicktekCommands functionality to interact with the firmware.

The device has a firmware version, normally SDKs are paired with forward
compantible firmwares but it is not guaranteed.

Signed-off-by: Sergio Abraham Martinez Rodriguez <sergio@flicktek.com>